feat: reestruturacao app
