+++ config.py
@@ -1,0 +1,13 @@
+""" Modulo Config """
+
+SECRET_KEY = 'alura'
+
+
+SQLALCHEMY_DATABASE_URI = \
+    '{SGBD}://{usuario}:{senha}@{servidor}/{database}'.format(
+        SGBD = 'mysql+mysqlconnector',
+        usuario = 'root',
+        senha = 'admin',
+        servidor = 'localhost',
+        database = 'jogoteca'
+    )
