+++ templates/lista.html
@@ -14,6 +14,10 @@
                     <td>{{ jogo.nome }}</td>
                     <td>{{ jogo.categoria }}</td>
                     <td>{{ jogo.console }}</td>
+                    <td>
+                        <a href="{{ url_for('editar', id=jogo.id) }}">Editar</a>
+                        <a href="{{ url_for('deletar', id=jogo.id) }}">Deletar</a>
+                    </td>
                 </tr>
                 {% endfor %}
